Switch to Emacs 24 package manager instead of our custom handling with git submodules
We manage to use the new package manager for Emacs 24 in all the packages that I use except for Rinari, as it depends on ruby-mode-1.1 which is not in the marmalade repo.  It is currently commented-out until I can figure things out.

Another is that since Emacs 24 contains it's own theme manager which does not have charcoal-black in it, I temporarily switched to wombat.  I don't like wombat, and I'll figure out how to get charcoal-black again!

Signed-off-by: Ealden Esto E. Escanan <ealden@gmail.com>